+++ src/js/interface.js
@@ -17,6 +17,15 @@ class Interface {
                 methodParameters: document.getElementById("current-method-parameters-container"),
                 methodReturns: document.getElementById("current-method-returns-container"),
                 methodVariables: document.getElementById("current-method-variables-container"),
+                methodPureCheckbox: document.getElementById("current-method-pure-checkbox"),
+            },
+            center: {
+                path: document.getElementById("path"),
+                methodName: document.getElementById("current-class-method-name"),
+                method: {
+                    container: document.getElementById("current-method"),
+                    nodeContainer: document.getElementById("node-container"),
+                }
             },
         };
     }
@@ -24,25 +33,27 @@ class Interface {
     refresh() {
         this.refreshLeft();
         this.refreshRight();
+        this.refreshCenter();
     }
 
     refreshLeft() {
         if (!this.app.currentClass) return;
-        this.elements.left.currentClassTitle.textContent = this.app.currentClass.name;
-        this.elements.left.currentMethodTitle.textContent = this.app.currentClass.currentMethod.name;
+        const left = this.elements.left;
+        left.currentClassTitle.textContent = this.app.currentClass.name;
+        left.currentMethodTitle.textContent = this.app.currentClass.currentMethod.name;
 
-        removeChildren(this.elements.left.classList);
-        removeChildren(this.elements.left.methodList);
-        removeChildren(this.elements.left.variableList);
-        removeChildren(this.elements.left.localVariableList);
+        removeChildren(left.classList);
+        removeChildren(left.methodList);
+        removeChildren(left.variableList);
+        removeChildren(left.localVariableList);
 
         this.refreshList({
             entries: Object.keys(this.app.classes),
             current: this.app.currentClass.name,
             exception: "Main",
-            parent: this.elements.left.classList,
+            parent: left.classList,
             calls: {
-                rename: (oldName, newName) => this.app.renameClass(oldName, newName),
+                change: (oldName, newName) => this.app.renameClass(oldName, newName),
                 click: (name) => this.app.openClass(name),
                 confirmRemove: {
                     message: (name) => `Are you sure you wan to remove class "${name}"?`,
@@ -60,9 +71,9 @@ class Interface {
             entries: Object.keys(this.app.currentClass.methods),
             current: this.app.currentClass.currentMethod.name,
             exception: "Constructor",
-            parent: this.elements.left.methodList,
+            parent: left.methodList,
             calls: {
-                rename: (oldName, newName) => this.app.currentClass.renameMethod(oldName, newName),
+                change: (oldName, newName) => this.app.currentClass.renameMethod(oldName, newName),
                 click: (name) => this.app.currentClass.openMethod(name),
                 confirmRemove: {
                     message: (name) => `Are you sure you wan to remove method "${name}" from ${this.app.currentClass.name}?`,
@@ -80,9 +91,9 @@ class Interface {
             entries: Object.keys(this.app.currentClass.variables),
             current: null,
             exception: null,
-            parent: this.elements.left.variableList,
+            parent: left.variableList,
             calls: {
-                rename: (oldName, newName) => this.app.currentClass.renameVariable(oldName, newName),
+                change: (oldName, newName) => this.app.currentClass.renameVariable(oldName, newName),
                 click: null,
                 confirmRemove: {
                     message: (name) => `Are you sure you wan to remove variable "${name}" from ${this.app.currentClass.name}?`,
@@ -100,9 +111,9 @@ class Interface {
             entries: Object.keys(this.app.currentClass.currentMethod.localVariables),
             current: null,
             exception: null,
-            parent: this.elements.left.localVariableList,
+            parent: left.localVariableList,
             calls: {
-                rename: (oldName, newName) => this.app.currentClass.currentMethod.renameLocalVariable(oldName, newName),
+                change: (oldName, newName) => this.app.currentClass.currentMethod.renameLocalVariable(oldName, newName),
                 click: null,
                 confirmRemove: {
                     message: (name) => `Are you sure you wan to remove local variable "${name}" from ${this.app.currentClass.currentMethod.name}?`,
@@ -121,11 +132,16 @@ class Interface {
     refreshRight() {
         if (!this.app.currentClass) return;
         const right = this.elements.right;
+
         removeChildren(right.classParameters);
         removeChildren(right.classVariables);
         removeChildren(right.methodParameters);
         removeChildren(right.methodReturns);
         removeChildren(right.methodVariables);
+        right.methodPureCheckbox.removeAttribute("disabled");
+
+        if (this.app.currentClass.currentMethod.name == "Constructor") right.methodPureCheckbox.setAttribute("disabled", "disabled");
+        right.methodPureCheckbox.checked = this.app.currentClass.currentMethod.pure;
 
         this.refreshList({
             entries: this.app.currentClass.parameters,
@@ -133,7 +149,7 @@ class Interface {
             exception: null,
             parent: right.classParameters,
             calls: {
-                rename: (oldName, newName) => this.app.currentClass.renameParameter(oldName, newName),
+                change: (oldName, newName) => this.app.currentClass.renameParameter(oldName, newName),
                 click: null,
                 confirmRemove: {
                     message: (name) => `Are you sure you wan to remove class parameter "${name}" from ${this.app.currentClass.name}?`,
@@ -153,7 +169,7 @@ class Interface {
             exception: null,
             parent: right.classVariables,
             calls: {
-                rename: (oldName, newName) => this.app.currentClass.setVariableValue(oldName, newName),
+                change: (oldName, newName) => this.app.currentClass.setVariableValue(oldName, newName),
                 click: null,
                 confirmRemove: null,
             },
@@ -165,7 +181,7 @@ class Interface {
             exception: null,
             parent: right.methodParameters,
             calls: {
-                rename: (oldName, newName) => this.app.currentClass.currentMethod.renameParameter(oldName, newName),
+                change: (oldName, newName) => this.app.currentClass.currentMethod.renameParameter(oldName, newName),
                 click: null,
                 confirmRemove: {
                     message: (name) => `Are you sure you wan to remove method parameter "${name}" from ${this.app.currentClass.currentMethod.name}?`,
@@ -179,9 +195,54 @@ class Interface {
                 },
             },
         });
+        this.refreshList({
+            entries: this.app.currentClass.currentMethod.returns,
+            current: null,
+            exception: null,
+            parent: right.methodReturns,
+            calls: {
+                change: (oldName, newName) => this.app.currentClass.currentMethod.renameReturn(oldName, newName),
+                click: null,
+                confirmRemove: {
+                    message: (name) => `Are you sure you wan to remove method return "${name}" from ${this.app.currentClass.currentMethod.name}?`,
+                    options: (name) => [
+                        { button: "Yes", call: () => {
+                            this.app.currentClass.currentMethod.removeReturn(name);
+                            this.app.success(`Successfully removed method return "${name}" from ${this.app.currentClass.currentMethod.name}`);
+                        }},
+                        { button: "Cancel", call: null },
+                    ],
+                },
+            },
+        });
+        this.refreshList({
+            entries: Object.entries(this.app.currentClass.currentMethod.localVariables),
+            current: null,
+            exception: null,
+            parent: right.methodVariables,
+            calls: {
+                change: (oldName, newName) => this.app.currentClass.currentMethod.setLocalVariableValue(oldName, newName),
+                click: null,
+                confirmRemove: null,
+            },
+        });
     }
 
-    refreshList({ entries, current, exception, parent, calls: { rename, click, confirmRemove } }) {
+    refreshCenter() {
+        if (!this.app.currentClass) return;
+        const center = this.elements.center;
+        center.methodName.textContent = this.app.currentClass.currentMethod.name;
+
+        removeChildren(center.path);
+
+        [this.app.currentClass.name, this.app.currentClass.currentMethod.name].forEach((t) => {
+            const title = document.createElement("p");
+            title.textContent = t;
+            center.path.appendChild(title);
+        });
+    }
+
+    refreshList({ entries, current, exception, parent, calls: { change, click, confirmRemove } }) {
         entries.forEach((entry) => {
             const isArray = Array.isArray(entry);
             const name = isArray ? entry[0] : entry; 
@@ -201,6 +262,7 @@ class Interface {
             input.value = isArray ? value : name;
             if (click) input.setAttribute("readonly", "readonly");
             if (current == name) input.classList.add("current");
+            if (exception == name) input.classList.add("read-only");
             li.appendChild(input);
 
             if (name != exception && click) input.addEventListener("dblclick", () => {
@@ -218,8 +280,8 @@ class Interface {
                 const tempOldName = entries[[...parent.children].map((li) => li.querySelector("input:not(:read-only)")).indexOf(input)];
                 if (!tempOldName) return;
                 const oldName = isArray ? tempOldName[0] : tempOldName;
-                input.setAttribute("readonly", "readonly");
-                if (rename && oldName != undefined) if (!rename(oldName, input.value)) input.value = oldName;
+                if (click) input.setAttribute("readonly", "readonly");
+                if (change && oldName != undefined) if (!change(oldName, input.value)) input.value = oldName;
             });
 
             if (click) li.addEventListener("click", (e) => {