+++ src/js/method.js
@@ -5,6 +5,8 @@ class Method {
         this.name = name;
 
         this.parameters = [];
+        this.returns = [];
+        this.pure = false;
         this.localVariables = {};
         this.uid = this.app.getNextUID(Object.values(this.class.methods).map((v) => v.uid));
     }
@@ -18,7 +20,7 @@ class Method {
     }
 
     createLocalVariable(name) {
-        this.app.manager.create(this.localVariables, name, "local variable", () => {
+        this.app.manager.create(this.localVariables, name, null, "local variable", () => {
             this.localVariables[name] = undefined;
         });
     }
@@ -32,7 +34,7 @@ class Method {
     }
 
     createParameter(name) {
-        this.app.manager.create(this.parameters, name, "method parameter", () => {
+        this.app.manager.create(this.parameters, name, null, "method parameter", () => {
             this.parameters.push(name);
         });    
     }
@@ -43,5 +45,30 @@ class Method {
 
     renameParameter(oldName, newName) {
         return this.app.manager.rename(this.parameters, oldName, newName, "Method parameter");
+    }
+
+    createReturn(name) {
+        this.app.manager.create(this.returns, name, null, "method return", () => {
+            this.returns.push(name);
+        });    
+    }
+
+    removeReturn(name) {
+        this.app.manager.remove(this.returns, name);
+    }
+
+    renameReturn(oldName, newName) {
+        return this.app.manager.rename(this.returns, oldName, newName, "Method return");
+    }
+
+    setLocalVariableValue(name, value) {
+        if (!Object.keys(this.localVariables).includes(name)) return this.app.error(`Local variable "${name}" does not exist in this method`);
+        this.localVariables[name] = value;
+        return true;
+    }
+
+    setPure(value) {
+        if (this.name == "Constructor") return;
+        this.pure = value;
     }
 };